services/api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
pydantic
